hockey-stats/requirements.txt
streamlit
pandas
numpy
plotly
streamlit-aggrid
psycopg2-binary
python-dotenv
